+++ agrobr/ibge/__init__.py
@@ -1,4 +1,4 @@
-"""Modulo IBGE - Dados PAM, LSPA, PPM, Abate, Censo Agropecuario, Censo Legado e Serie Historica."""
+"""Modulo IBGE - Dados PAM, LSPA, PPM, Abate, Censo Agropecuario, Censo Legado, Serie Historica e Municipal 1985."""
 
 from __future__ import annotations
 
@@ -17,6 +17,10 @@ from agrobr.ibge.api import (
     temas_censo_agro_historico,
     ufs,
 )
+from agrobr.ibge.censo_municipal_1985 import (
+    censo_agro_municipal_1985,
+    temas_censo_agro_municipal_1985,
+)
 from agrobr.ibge.legacy_api import censo_agro_legado, temas_censo_agro_legado
 
 __all__ = [
@@ -24,6 +28,7 @@ __all__ = [
     "censo_agro",
     "censo_agro_historico",
     "censo_agro_legado",
+    "censo_agro_municipal_1985",
     "especies_abate",
     "especies_ppm",
     "lspa",
@@ -34,5 +39,6 @@ __all__ = [
     "temas_censo_agro",
     "temas_censo_agro_historico",
     "temas_censo_agro_legado",
+    "temas_censo_agro_municipal_1985",
     "ufs",
 ]